Modulo16Lab/project-mgmt-fastapi/backend/requirements.txt
fastapi
uvicorn
mysql-connector-pythonx